+++ strucmotif-search-core/src/main/java/org/rcsb/strucmotif/domain/query/MotifSearchQuery.java
@@ -36,7 +36,7 @@ public class MotifSearchQuery {
     private final Map<LabelSelection, Set<ResidueType>> exchanges;
     private final Collection<String> whitelist;
     private final Collection<String> blacklist;
-    private final SearchSpace searchSpace;
+    private final StructureDeterminationMethodology structureDeterminationMethodology;
     private final MotifSearchConfig config;
 
     MotifSearchQuery(MotifSearchRuntime motifSearchRuntime,
@@ -48,7 +48,7 @@ public class MotifSearchQuery {
                      Map<LabelSelection, Set<ResidueType>> exchanges,
                      Collection<String> whitelist,
                      Collection<String> blacklist,
-                     SearchSpace searchSpace,
+                     StructureDeterminationMethodology structureDeterminationMethodology,
                      MotifSearchConfig config) {
         this.motifSearchRuntime = motifSearchRuntime;
         List<ResiduePairOccurrence> residuePairOccurrences = parameters.getMotifPruner().prune(structure, labelSelections, residues);
@@ -57,7 +57,7 @@ public class MotifSearchQuery {
         this.exchanges = exchanges;
         this.whitelist = whitelist;
         this.blacklist = blacklist;
-        this.searchSpace = searchSpace;
+        this.structureDeterminationMethodology = structureDeterminationMethodology;
         this.config = config;
     }
 
@@ -105,8 +105,8 @@ public class MotifSearchQuery {
      * Access to the applied search space.
      * @return ALL, PDB, or MODELS
      */
-    public SearchSpace getSearchSpace() {
-        return searchSpace;
+    public StructureDeterminationMethodology getSearchSpace() {
+        return structureDeterminationMethodology;
     }
 
     /**